[AvrTemp] Reimplement some logic and write some critical tests

+++ src/YellowBelt/Week1/avrTemp/avrTemp.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-int64_t calculateAverage(const vector<int> &data) {
+int64_t calculateAverage(const vector<int64_t> &data) {
     int64_t sum = 0;
     for (const auto &item : data) {
         sum += item;
@@ -13,7 +13,7 @@ int64_t calculateAverage(const vector<int> &data) {
     return sum / elementCount;
 }
 
-vector<int> getIndexesHigherThanValue(const vector<int> &data, const int64_t &value) {
+vector<int> getIndexesHigherThanValue(const vector<int64_t> &data, const int64_t &value) {
     vector<int> result;
 
     for (int i = 0; i < data.size(); ++i) {
@@ -25,7 +25,7 @@ vector<int> getIndexesHigherThanValue(const vector<int> &data, const int64_t &va
     return result;
 }
 
-vector<int> process(const vector<int> &temperatures) {
+vector<int> process(const vector<int64_t> &temperatures) {
     int64_t average = calculateAverage(temperatures);
     return getIndexesHigherThanValue(temperatures, average);
 }
