+++ src/YellowBelt/Week1/avrTemp/avrTemp.h
@@ -1,5 +1,6 @@
 #include <vector>
+#include <cstdint>
 
 using namespace std;
 
-vector<int> process(const vector<int> &temperatures);
+vector<int> process(const vector<int64_t> &temperatures);
