+++ src/YellowBelt/Week1/avrTemp/avrTempCoursera.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-int64_t calculateAverage(const vector<int> &data) {
+int64_t calculateAverage(const vector<int64_t> &data) {
     int64_t sum = 0;
     for (const auto &item : data) {
         sum += item;
@@ -14,7 +14,7 @@ int64_t calculateAverage(const vector<int> &data) {
     return sum / elementCount;
 }
 
-vector<int> getIndexesHigherThanValue(const vector<int> &data, const int64_t &value) {
+vector<int> getIndexesHigherThanValue(const vector<int64_t> &data, const int64_t &value) {
     vector<int> result;
 
     for (int i = 0; i < data.size(); ++i) {
@@ -26,7 +26,7 @@ vector<int> getIndexesHigherThanValue(const vector<int> &data, const int64_t &va
     return result;
 }
 
-vector<int> process(const vector<int> &temperatures) {
+vector<int> process(const vector<int64_t> &temperatures) {
     int64_t average = calculateAverage(temperatures);
     return getIndexesHigherThanValue(temperatures, average);
 }
@@ -36,7 +36,7 @@ int main() {
     size_t values_count;
     cin >> values_count;
 
-    vector<int> input_data;
+    vector<int64_t> input_data;
 
     int temp;
     for (size_t i = 0; i < values_count; ++i) {
